+++ app/src/main/java/com/facebook/profilo/sample/model/Point.java
@@ -4,7 +4,7 @@ public class Point extends com.facebook.profilo.sample.model.ttypes.Point {
     private Trace trace;
     private ExecutionUnit unit;
     private Block block;
-    public Point(Trace trace, ExecutionUnit unit, Block block, int timestamp, int sequenceNumber)
+    public Point(Trace trace, ExecutionUnit unit, Block block, long timestamp, int sequenceNumber)
     {
         super(GenID.new_id(),timestamp,timestamp,new Properties(),sequenceNumber);
         this.trace = trace;
@@ -15,7 +15,7 @@ public class Point extends com.facebook.profilo.sample.model.ttypes.Point {
     {
         return this.id;
     }
-    public int getTimestamp()
+    public long getTimestamp()
     {
         return this.timestamp;
     }
